3d_extractor_JJ/RAD/requirements.txt
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

matplotlib==3.9.4
numpy==2.2.6
opencv_python_headless==4.6.0.66
pandas==2.2.3
Pillow==10.4.0
scikit_image==0.24.0
scikit_learn==1.5.2
scipy==1.14.1
tabulate==0.9.0
torch==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.64.1
ptflops==0.7
timm==0.9.12
open3d
tifffile
